# backend/tic-tac-toe/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      message:
        description: Error message.
        example: invalid request body
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      login:
        description: User login.
        example: player
        type: string
      uuid:
        description: User UUID.
        example: 123e4567-e89b-42d3-a456-426614174000
        format: uuid
        type: string
    type: object
info:
  contact: {}
  description: HTTP API for user authentication and tic-tac-toe games.
  title: Tic-Tac-Toe API
  version: "1.0"
paths:
...
  /users/{uuid}:
    get:
      description: Returns public user information. Request body is not used.
      parameters:
      - default: 123e4567-e89b-42d3-a456-426614174000
        description: User UUID
        format: uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Missing or invalid Bearer token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: User was not loaded
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by UUID
      tags:
      - users
  /users/me:
    delete:
      description: Soft-deletes the authenticated user account by marking it deleted
        and making the login unavailable for future auth.
      produces:
      - application/json
      responses:
        "204":
          description: User deleted
        "401":
          description: Missing or invalid Bearer token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: User was not deleted
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete current user
      tags:
      - users
    get:
      description: Returns public user information for the authenticated Bearer token.
      produces:
      - application/json
      responses:
        "200":
          description: User
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Missing or invalid Bearer token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: User was not loaded
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"